Add solutions to codeforces problems
Solved 'Anton and Letters' and 'Fox and Snakes'. Had
a bit of a hiccup on 'Anton and Letters'. It turns
out that for inputs with a whitespace I should
use the getline() function. My logic is correct,
the input just messed me up. So I had to solve
another problem to bring back my morale which is
'Fox and Snakes'. It's a simple draw the pattern
problem that requires a bit of thinking.

Signed-off-by: Christopher Quinto <cquinto.primary@gmail.com>